reverted to old redux, created seperate branch for redux update
