--- src/API/ChatRooms/CreateNewRoom.js
@@ -1,10 +1,0 @@
-import { API } from '../axios'
-
-export async function createNewRoomAPI(data) {
-    try {
-        const response = await API.post('/rooms/new', data);
-        return response;
-    } catch (err) {
-        throw err;
-    }
-}
