--- src/API/ChatRooms/GetAllRooms.js
@@ -1,10 +1,0 @@
-import { API } from '../axios'
-
-export async function getAllRoomsAPI() {
-    try {
-        const response = await API.get('/rooms/all');
-        return response;
-    } catch (err) {
-        throw err;
-    }
-}
