--- src/API/ChatRooms/index.js
@@ -1,4 +1,0 @@
-export * from './CreateNewRoom';
-export * from './GetAllRooms';
-export * from './GetChatRoom';
-export * from './SubmitMessage';
