--- src/API/axios.js
@@ -1,5 +1,0 @@
-import axios from 'axios'
-
-export const API = axios.create({
-	baseURL: process.env.REACT_APP_API_URL
-})
