--- src/API/index.js
@@ -1,2 +1,0 @@
-export * from './ChatRooms';
-export * from './User';
