+++ src/Components/Chatbar/index.js
@@ -1,26 +1,19 @@
 import React, { useState, useEffect } from 'react';
 import { ChatBarStyles } from './Styles';
-
 import { ChatHeader, ChatBody, ChatFooter } from './Components'
-import { useSelector, useDispatch } from 'react-redux'
-import { GetChatRoomThunk } from '../../Redux/ReduxStore'
 
-import { useStateValue } from "../../StateProvider";
-import axios from '../../axios';
+import { useStateValue } from "../../Redux/StateProvider";
+import axios from '../../Redux/axios';
 
-export function ChatBar() {
+export function ChatBar({ messages, currentRoom }) {
     const classes = ChatBarStyles();
-    const dispatch = useDispatch();
-    const { action, currentRoom, currentMessages } = useSelector((state) => state.ReduxStore);
 
     const [input, setInput] = useState('');
 	const [{ user }] = useStateValue();
 	const [seed, setSeed] = useState('');
 	
-    useEffect(async () => {
+    useEffect(() => {
 		setSeed(Math.floor(Math.random() * 5000))
-        await dispatch(GetChatRoomThunk({room_id: currentRoom._id}));
-        await console.log(action)
 	}, []);
 
     // input on change handle
@@ -54,7 +47,7 @@ export function ChatBar() {
                 timestamp={currentRoom.lastTimestamp}
             />
             <ChatBody 
-                messages={currentMessages}
+                messages={messages}
                 username={user.displayName}
             />
             <ChatFooter
