+++ src/Redux/StateProvider.js
@@ -1,11 +1,11 @@
-// import React, { createContext, useContext, useReducer } from "react";
+import React, { createContext, useContext, useReducer } from "react";
 
-// export const StateContext = createContext();
+export const StateContext = createContext();
 
-// export const StateProvider = ({ reducer, initialState, children }) => (
-// 	<StateContext.Provider value={useReducer(reducer, initialState)}>
-// 		{children}
-// 	</StateContext.Provider>
-// );
+export const StateProvider = ({ reducer, initialState, children }) => (
+	<StateContext.Provider value={useReducer(reducer, initialState)}>
+		{children}
+	</StateContext.Provider>
+);
 
-// export const useStateValue = () => useContext(StateContext);
+export const useStateValue = () => useContext(StateContext);
